Move UI-related files to the presentation package
