+++ app/src/main/java/com/artemklymenko/mychat/MainActivity.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.artemklymenko.mychat.navigation.RootAppNavigation
-import com.artemklymenko.mychat.ui.theme.MyChatTheme
+import com.artemklymenko.mychat.presentation.navigation.RootAppNavigation
+import com.artemklymenko.mychat.presentation.ui.theme.MyChatTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
