+++ app/src/main/java/com/artemklymenko/mychat/presentation/pages/main/MainScreen.kt
@@ -1,0 +1,9 @@
+package com.artemklymenko.mychat.presentation.pages.main
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun MainScreen(modifier: Modifier) {
+
+}
